# aps/nodejs/angular-ui/src/users/dto/update-user.dto.ts
import { IsOptional, IsString, IsNotEmpty, MinLength, IsArray, IsEmail } from 'class-validator';

export class UpdateUserDto {
  @IsString()
  @IsNotEmpty()
  @MinLength(4)
  username: string;

  @IsString()
  @IsNotEmpty()
  firstName: string;

  @IsString()
  @IsNotEmpty()
  lastName: string;

  @IsString()
  @IsEmail()
  @IsNotEmpty()
  email: string;

  @IsOptional()
  @IsString()
  @MinLength(6)
  password: string;

  @IsArray()
  roles: string[];
}