# aps/nodejs/angular-ui/swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_root","parameters":[],"responses":{"200":{"description":""}}}},"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/login":{"get":{"operationId":"AppController_index","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"AppController_login","parameters":[],"responses":{"201":{"description":""}}}},"/profile":{"get":{"operationId":"AppController_protected","parameters":[],"responses":{"200":{"description":""}}}},"/logout":{"get":{"operationId":"AppController_logout","parameters":[],"responses":{"200":{"description":""}}}},"/auth/login":{"post":{"operationId":"AuthController_apiLogin","parameters":[],"responses":{"201":{"description":""}}}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}}}},"/api/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/api/users/{id}":{"post":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/users/me":{"get":{"operationId":"UsersController_findMe","parameters":[],"responses":{"200":{"description":""}}}},"/api/vms":{"post":{"operationId":"VmsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUpdateVmsDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"VmsController_findAll","parameters":[{"name":"with_vms_info","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/vms/{id}":{"post":{"operationId":"VmsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUpdateVmsDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"VmsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"with_vms_info","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"VmsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/vms/{id}/devices":{"post":{"operationId":"VmsController_createDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUpdateDeviceDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"VmsController_findAllDevices","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/vms/{id}/devices/{deviceId}":{"post":{"operationId":"VmsController_updateDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUpdateDeviceDto"}}}},"responses":{"201":{"description":""}}},"get":{"operationId":"VmsController_findOneDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"VmsController_removeDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/vms/{id}/devices/{deviceId}/measurements":{"get":{"operationId":"VmsController_findMeasurementsForDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/vms/{id}/devices/{deviceId}/lastmeasurement":{"get":{"operationId":"VmsController_findLastMeasurementsForDevice","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}},{"name":"deviceId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/api/exchange/{din}/bind":{"post":{"operationId":"ExchangeController_bind","parameters":[{"name":"din","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/api/exchange/{din}/unbind":{"post":{"operationId":"ExchangeController_unbind","parameters":[{"name":"din","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/api/exchange/{din}/messages":{"post":{"operationId":"ExchangeController_sendMessage","parameters":[{"name":"din","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}}}},"/api/exchange/recorders":{"get":{"operationId":"ExchangeController_getRecorders","parameters":[],"responses":{"200":{"description":""}}}},"/api/exchange/vms-list":{"get":{"operationId":"ExchangeController_getVmsList","parameters":[],"responses":{"200":{"description":""}}},"post":{"operationId":"ExchangeController_refreshVmsList","parameters":[],"responses":{"201":{"description":""}}}},"/api/exchange/notifications":{"post":{"operationId":"WebHookController_vmsNotification","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebHookDto"}}}},"responses":{"201":{"description":""}}}}},"info":{"title":"Daas Application","description":"API Documentation","version":"v1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"CreateUpdateVmsDto":{"type":"object","properties":{"name":{"type":"string"},"iot":{"type":"boolean"},"vmsUri":{"type":"string"},"vmsPort":{"type":"number"},"note":{"type":"string"}},"required":["name","iot","vmsUri","vmsPort","note"]},"CreateUpdateDeviceDto":{"type":"object","properties":{"din":{"type":"string"},"name":{"type":"string"},"typeset":{"type":"string"},"note":{"type":"string"}},"required":["din","name","typeset","note"]},"WebHookDto":{"type":"object","properties":{}}}}}